feat: Add spinners for animations
This commit adds a new file `animate.py` which contains a function `spinner` that displays spinning characters for a specified duration. The function uses a list of spin characters and updates the stdout by overwriting the previous character. The duration of the spinner animation can be specified by changing the argument passed to the function.